Refactor imports to streamline module dependencies and remove core references
