+++ src/.env-example
@@ -3,7 +3,8 @@ SHARDS=1
 version=2.1.0
 MODE=development
 LOCAL_LAVA=false
-
+X_CLIENT_ID=
+X_CLIENT_SECRET=
 GEMINI_KEY=gemini_api_key
 MONGO_URI='mongodb://localhost:27017/db'
 BOT_TOKEN=your_bot_token
