+++ src/cogs/esports/scrim.py
@@ -13,7 +13,7 @@ from ext import constants, checks
 from discord.ext import commands, tasks
 from datetime import datetime, timedelta
 from ext.models.scrim import ScrimModel, Team
-from core.config import IS_DEV_ENV
+from config import IS_DEV_ENV
 from discord import Embed, TextChannel,  Interaction,   app_commands as app
